+++ source/rafcon/gui/controllers/right_click_menu/state.py
@@ -142,9 +142,10 @@ class StateMachineRightClickMenu(object):
         if len(selection.states) == 1 and not selected_state_m.state.is_root_state:
             save_as_sub_menu_item, save_as_sub_menu = append_sub_menu_to_parent_menu("Save state as", menu,
                                                                                      constants.BUTTON_SAVE)
-
+            callback_function = partial(self.on_save_as_activate,
+                                        save_as_function=gui_helper_state_machine.save_selected_state_as)
             save_as_sub_menu.append(create_image_menu_item("State machine", constants.BUTTON_SAVE,
-                                                           self.on_save_state_as_state_machine_activate,
+                                                           callback_function,
                                                            accel_code=shortcuts_dict['save_state_as'][0],
                                                            accel_group=accel_group))
             save_as_library_sub_menu_item, save_as_library_sub_menu = append_sub_menu_to_parent_menu("Library",
@@ -152,9 +153,32 @@ class StateMachineRightClickMenu(object):
                                                                                                      constants.SIGN_LIB)
             library_root_paths = core_singletons.library_manager.library_root_paths
             for library_root_key in library_root_paths.iterkeys():
+                callback_function = partial(self.on_save_as_activate,
+                                            path=library_root_paths[library_root_key],
+                                            save_as_function=gui_helper_state_machine.save_selected_state_as)
                 save_as_library_sub_menu.append(create_image_menu_item(library_root_key, constants.SIGN_LIB,
-                                                                       partial(self.on_save_state_as_state_machine_activate,
-                                                                               path=library_root_paths[library_root_key]),
+                                                                       callback_function,
+                                                                       accel_code=None, accel_group=accel_group))
+        else:
+            save_as_sub_menu_item, save_as_sub_menu = append_sub_menu_to_parent_menu("Save state machine as", menu,
+                                                                                     constants.BUTTON_SAVE)
+
+            callback_function = partial(self.on_save_as_activate,
+                                        save_as_function=gui_helper_state_machine.save_state_machine_as)
+            save_as_sub_menu.append(create_image_menu_item("State machine", constants.BUTTON_SAVE,
+                                                           callback_function,
+                                                           accel_code=shortcuts_dict['save_as'][0],
+                                                           accel_group=accel_group))
+            save_as_library_sub_menu_item, save_as_library_sub_menu = append_sub_menu_to_parent_menu("Library",
+                                                                                                     save_as_sub_menu,
+                                                                                                     constants.SIGN_LIB)
+            library_root_paths = core_singletons.library_manager.library_root_paths
+            for library_root_key in library_root_paths.iterkeys():
+                callback_function = partial(self.on_save_as_activate,
+                                            path=library_root_paths[library_root_key],
+                                            save_as_function=gui_helper_state_machine.save_state_machine_as)
+                save_as_library_sub_menu.append(create_image_menu_item(library_root_key, constants.SIGN_LIB,
+                                                                       callback_function,
                                                                        accel_code=None, accel_group=accel_group))
 
         return menu
@@ -294,15 +318,17 @@ class StateMachineRightClickMenu(object):
         self.shortcut_manager.trigger_action('run_to_selected', None, None)
 
     @staticmethod
-    def on_save_state_as_state_machine_activate(widget, data=None, path=None):
+    def on_save_as_activate(widget, data=None, path=None, save_as_function=None):
         # workaround to set the initial path in the 'choose folder' dialog to the handed one
         old_last_path_open = gui_singletons.global_runtime_config.get_config_value('LAST_PATH_OPEN_SAVE', None)
+        if save_as_function is None:
+            logger.error("Hand a function for operation 'save_as' currently it is '{0}'".format(save_as_function))
+            return
+
         try:
             if path is not None:
                 gui_singletons.global_runtime_config.set_config_value('LAST_PATH_OPEN_SAVE', path)
-                gui_helper_state_machine.save_selected_state_as()
-            else:
-                gui_helper_state_machine.save_selected_state_as()
+            save_as_function()
         except Exception:
             raise
         finally: